+++ test/hoster_http_test.go
@@ -1,4 +1,4 @@
-package gohost
+package test
 
 import (
 	"crypto/tls"
